Remove reflection impl to rely on glue impl
