+++ lib/build.gradle
@@ -57,7 +57,7 @@ dependencies {
 	palantirJavaFormatCompileOnly 'com.palantir.javaformat:palantir-java-format:1.1.0' 	// this version needs to stay compilable against Java 8 for CI Job testNpm
 
 	// used jackson-based formatters
-	jacksonCompileOnly 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.13.4'
+	jacksonCompileOnly 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.14.1'
 
 	String VER_KTFMT = '0.42'
 	ktfmtCompileOnly "com.facebook:ktfmt:$VER_KTFMT"
