--- lib/src/main/java/com/diffplug/spotless/yaml/YamlJacksonV2Step.java
@@ -1,77 +1,0 @@
-/*
- * Copyright 2021-2023 DiffPlug
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.diffplug.spotless.yaml;
-
-import java.io.IOException;
-import java.io.Serializable;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-
-import com.diffplug.spotless.FormatterFunc;
-import com.diffplug.spotless.FormatterStep;
-import com.diffplug.spotless.JarState;
-import com.diffplug.spotless.Provisioner;
-
-public class YamlJacksonV2Step {
-	private YamlJacksonV2Step() {}
-
-	public static String defaultVersion() {
-		return YamlJacksonStep.DEFAULT_VERSION;
-	}
-
-	public static FormatterStep create(List<String> enabledFeatures,
-			List<String> disabledFeatures,
-			String jacksonVersion,
-			Provisioner provisioner) {
-		Objects.requireNonNull(provisioner, "provisioner cannot be null");
-		return FormatterStep.createLazy("yaml",
-				() -> new State(enabledFeatures, disabledFeatures, jacksonVersion, provisioner),
-				State::toFormatter);
-	}
-
-	public static FormatterStep create(Provisioner provisioner) {
-		return create(Arrays.asList("INDENT_OUTPUT"), Arrays.asList(), defaultVersion(), provisioner);
-	}
-
-	private static final class State implements Serializable {
-		private static final long serialVersionUID = 1L;
-
-		private final List<String> enabledFeatures;
-		private final List<String> disabledFeatures;
-
-		private final JarState jarState;
-
-		private State(List<String> enabledFeatures,
-				List<String> disabledFeatures,
-				String jacksonVersion,
-				Provisioner provisioner) throws IOException {
-			this.enabledFeatures = enabledFeatures;
-			this.disabledFeatures = disabledFeatures;
-
-			this.jarState = JarState.from(YamlJacksonStep.MAVEN_COORDINATE + jacksonVersion, provisioner);
-		}
-
-		FormatterFunc toFormatter() throws ClassNotFoundException, NoSuchMethodException, InvocationTargetException,
-				InstantiationException, IllegalAccessException {
-			Class<?> formatterFunc = jarState.getClassLoader().loadClass("com.diffplug.spotless.glue.yaml.YamlJacksonFormatterFunc");
-			Constructor<?> constructor = formatterFunc.getConstructor(List.class, List.class);
-			return (FormatterFunc) constructor.newInstance(enabledFeatures, disabledFeatures);
-		}
-	}
-}
